refactor `AnalysisInputLocation`s `Collection getClassSources(...)` to `Stream getClassSources(...) -> passing the Autoclosable object to the callers to avoid collecting streams and re-stream() them in the meantime
