+++ sootup.java.bytecode.frontend/src/test/java/sootup/java/bytecode/frontend/Soot1577Test.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 import sootup.core.inputlocation.AnalysisInputLocation;
 import sootup.core.model.SootMethod;
 import sootup.core.model.SourceType;
-import sootup.java.bytecode.frontend.inputlocation.PathBasedAnalysisInputLocation;
+import sootup.java.bytecode.frontend.inputlocation.ClassFileBasedAnalysisInputLocation;
 import sootup.java.core.views.JavaView;
 
 @Tag(TestCategories.JAVA_8_CATEGORY)
@@ -19,7 +19,7 @@ public class Soot1577Test {
   @Test
   public void test() {
     AnalysisInputLocation inputLocation =
-        new PathBasedAnalysisInputLocation.ClassFileBasedAnalysisInputLocation(
+        new ClassFileBasedAnalysisInputLocation(
             Paths.get("../shared-test-resources/soot-1577/g.class"),
             "cn.com.chinatelecom.account.api.c",
             SourceType.Application);
